feat: SNS, Avatar 클래스 전부 BusinessCard 파일로 통합
