+++ app/src/main/java/com/hongul/filq/model/BusinessCard.kt
@@ -1,5 +1,8 @@
 package com.hongul.filq.model
 
+import androidx.annotation.DrawableRes
+import com.hongul.filq.R
+
 data class BusinessCard(
     // 명함 고유 id(userId@prefix(autoIncremented) 형식)
     // ex) 94842789402141@24
@@ -19,3 +22,46 @@ data class BusinessCard(
     val avatar: Avatar, // 명함 프로필 이미지
     val introduction: String, // 명함 소개
 )
+
+sealed class SNS(
+    @DrawableRes val icon: Int,
+    val link: String,
+    val idParser: (String) -> String
+) {
+    val userId: String
+        get() = idParser(link)
+
+    class Facebook(link: String) :  SNS(R.drawable.logo_facebook, link, DEFAULT_PARSER)
+    class Instagram(link: String) : SNS(R.drawable.logo_instagram, link, DEFAULT_PARSER)
+    class X(link: String) :         SNS(R.drawable.logo_x, link, DEFAULT_PARSER)
+    class Youtube(link: String) :   SNS(R.drawable.logo_youtube, link, DEFAULT_PARSER)
+
+    companion object {
+        val DEFAULT_PARSER: (String) -> String = { link -> link.split("/").last() }
+    }
+}
+
+data class Avatar(
+    val path: String? = null,
+    val sticker: Sticker? = null
+) {
+    val isImage: Boolean
+        get() = path != null
+    val isSticker: Boolean
+        get() = sticker != null
+    val isDefault: Boolean
+        get() = !isImage && !isSticker
+}
+
+data class Sticker(
+    val pos: Int,
+    val color: Long
+) {
+    companion object {
+        const val SIZE = 256
+    }
+
+    val size: Int get() = SIZE
+    val x: Int get() = (pos % 4) * size
+    val y: Int get() = (pos / 4) * size
+}
